requirements.txt
azure-core==1.29.3
azure-cosmos==4.5.0
azure-functions==1.17.0
azure-identity==1.14.0
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
isodate==0.6.1
msal==1.23.0
msal-extensions==1.0.0
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
pywin32==306
requests==2.31.0
six==1.16.0
typing-extensions==4.7.1
urllib3==2.0.4
